backend/requirements.txt
django>=5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
scikit-learn>=1.4
pandas>=2.2
numpy>=1.26
joblib>=1.3
google-generativeai>=0.8
